More Inkopod Class
States "egg", "idle" and "busy".
Egg must wait a while before it can hatch to an idle state.
Idle state is contributing to production (gain food) and reproduction (make babies!)
Busy state currently does nothing, but is not an egg and is not idle.

Eggs use "age" and "mature" to determine when it may be hatched.

Special type "Wild" is a mystery egg, that randomly becomes a different type when hatched.

Removed private fields, it's making me too crazy.