Saving the Population, Create and Restore Methods
save and load now functioning to save all the inkopods players collect, and load them back on page reload.

The constructor for the inkopod class is now effectively private, since it can only be utilized with a couple of static methods - create and restore.
Restore is intended to be used when loading the game.